+++ CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 Planned changes/updates/fixes/features for to be part of one of the next _releases/versions_
 
-## [0.6.0-SNAPSHOT] - 2019-01-xx
+## [0.6.0] - 2019-01-29
 ### Added
   * Support for creation of absolute path with not existing parent directories in FTP Server
     * In such cases all not existing path elements are automatically created